Initial Project Files Upload
4-6-2023

Thomas Smith uploaded the initial project files for the behavior application along with necessary folders and accessory code. Customizations to 'BehaviorApp.m' still need to be generalized for the public and authors need to be listed within code text. Lastly, a tutorial/manual needs to be created for practical use.